% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffHub.R
\name{DiffPR.HCNetlas}
\alias{DiffPR.HCNetlas}
\title{DiffPR.HCNetlas}
\usage{
DiffPR.HCNetlas(rank.df.final = NULL, control = NULL, disease = NULL)
}
\arguments{
\item{rank.df.final}{Output of CombinePercRank}

\item{control}{Character string that states disease}

\item{disease}{Character string that states control}
}
\value{
Dataframe of celltypes and their diffPR values of each genes in the CGNs
}
\description{
Get differnence of normalized centrality values from the output of GetCentrality()
}
\examples{
strength.list <- GetCentrality(net.list = merged.net.list)
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
DiffPR.HCNetlas(rank.df.final, control = 'HCNetlas', disease = 'SLE')
}
