% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffHub.R
\name{FindDiffHub.HCNetlas}
\alias{FindDiffHub.HCNetlas}
\title{FindDiffHub.HCNetlas}
\usage{
FindDiffHub.HCNetlas(
  rank.df.final = NULL,
  disease = NULL,
  control = NULL,
  net.list = NULL,
  centrality = "degree",
  q.method = "BH",
  min.cells = 1000,
  dis.meta = NULL,
  celltypes = NULL
)
}
\arguments{
\item{rank.df.final}{Output from CombinePercRank}

\item{disease}{Character string that states control}

\item{control}{Character string that states disease}

\item{net.list}{Output of SortAddLLS, network list to perfrom the analysis}

\item{centrality}{the centrality method used in the getCentrality. both must match! this is used to find the null distribution}

\item{q.method}{Method to perfrom multiple testing correction. accepted values are: c("BH","holm","hochberg","bonferroni","BY","fdr","none")}

\item{min.cells}{minimum cells required to perform diff network analysis default value is 500}

\item{dis.meta}{disease metadata data.frame used for this analayis}

\item{celltypes}{column name where celltypes annotaiton are stored}
}
\value{
Dataframe Percentile rank of centrality of each networks, their difference, and their statistical significance
}
\description{
Calcuate statistical signficance of diffPR values calculated with DiffPR()
}
\examples{
strength.list <- GetCentrality(net.list = merged.net.list)
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
diffPR.df.sig <- FindDiffHub.HCNetlas(rank.df.final, control = 'HCNetlas', disease='SLE', net.list = merged.net.list, centrality = "degree", q.method = "BH", min.cells = 1000, dis.meta = meta_select, celltypes = "celltypes_merged")
}
