% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesetConnectivity.R
\name{GenesetConnectivity}
\alias{GenesetConnectivity}
\title{GenesetConnectivity}
\usage{
GenesetConnectivity(
  net.list = NULL,
  control = NULL,
  disease = NULL,
  geneset = NULL
)
}
\arguments{
\item{net.list}{Output of SortAddLLS, network list to perfrom the analysis}

\item{control}{Character string that states disease}

\item{disease}{Character string that states control}

\item{geneset}{Character string that contains disease-related genes}
}
\value{
Dataframe Count of connectivitiy in each networks
}
\description{
Calcuate conncectivity among disease genes in each networks
}
\examples{
geneset.conn <- GenesetConnectivity(net.list = merged.net.list, control = 'hcNETLAS', disease = 'SLE', geneset = sle_genes)
}
