% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcNETLAS_base.R
\name{getCelltype}
\alias{getCelltype}
\title{getCelltype}
\usage{
getCelltype(tissue = NULL)
}
\arguments{
\item{tissue}{a character of tissue names. Should be included in output of Tissues() function.}
}
\value{
a vector of characters, with cell-type names.
}
\description{
Return cell types of networks stored in hcNETLAS in a specific tissue.
}
\examples{
getCelltype(tissue="LNG")
}
