% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopHubGene.R
\name{GetNeighbors}
\alias{GetNeighbors}
\title{GetNeighbors}
\usage{
GetNeighbors(top.df = NULL, net.list = NULL)
}
\arguments{
\item{top.df}{Output of TopHub, a dataframe with top hubness genes}

\item{net.list}{List of networks. Should contain all networks from top.df}
}
\value{
List Characters of direct neighboring nodes and top hub genes in each network
}
\description{
Find direct neibhbors for each CGNs in corresponding top hub genes
}
\examples{
direct.neighbors <- GetNeighbors(top.df=sel_top.df, net.list =merged.net.list)
}
