% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopHubGene.R
\name{GSAplot}
\alias{GSAplot}
\title{GSAplot}
\usage{
GSAplot(genes, database, title, top.term)
}
\arguments{
\item{genes}{Vector a gene set of interest}

\item{database}{Character indicating which pathway to use. Should be db contained in EnrichR}

\item{title}{Title label of GSEA plot}

\item{top.term}{Number of terms that will be represented in a plot}
}
\value{
Plot with top n pathways enriched in a given gene set.
}
\description{
Plot Gene set enrichment plot using EnrichR
}
\examples{
go <- GSAplot(top.genes,'GO_Biological_Process_2021',"SLE_Bcell_gene GOBP",10)
}
