% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcNETLAS_base.R
\name{MergeCGN.hcNETLAS}
\alias{MergeCGN.hcNETLAS}
\title{MergeCGN.hcNETLAS}
\usage{
MergeCGN.hcNETLAS(disease_cgn = NULL, disease = NULL, tissue = NULL)
}
\arguments{
\item{disease_cgn}{scHumanNet output CGN list. output from SortAddLLS() function in scHumanNet package.}

\item{disease}{a character specifying a disease.}

\item{tissue}{a tissue name for selecting hcNETLAS CGNs. We recomand this corresponds with tissue where disease scRNA-seq data were obtained. Should be included in output of Tissues() function.}
}
\value{
a list of CGN networks. Disease CGNs and hcNETLAS healhty CGNs merged.
}
\description{
Merge disease CGNs and hcNETLAS CGNs with corresponding tissue.
}
\examples{
MergeCGN.hcNETLAS(disease_cgn=sorted.net.list, disease="SLE", tissue="LNG")
}
