% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcNETLAS_base.R
\name{MergeCGN}
\alias{MergeCGN}
\title{MergeCGN}
\usage{
MergeCGN(
  disease_cgn = NULL,
  disease = NULL,
  control_cgn = NULL,
  control = NULL
)
}
\arguments{
\item{disease_cgn}{a list of disease CGNs. scHumanNet output from SortAddLLS() function in scHumanNet package.}

\item{disease}{a character specifying a disease.}

\item{control_cgn}{a list of healthy CGNs. scHumanNet output from SortAddLLS() function in scHumanNet package.}

\item{control}{a character specifying control.}
}
\value{
a list of CGN networks. Disease CGNs and healhty CGNs merged.
}
\description{
Merge disease CGNs and healthy CGNs to generate list of networks.
}
\examples{
MergeCGN(disease_cgn=dis.sorted.net.list, disease="SLE", control_cgn=ctl.sorted.net.list, control="Control")
}
